Fix a theoretical bug where ColumnText.add() could fail to add a string.
I think this is only theoretical, visible in unit tests, because the construction of a metadata page in Sas7bdatWriter places the ColumnTextSubheader in such a way as to make this impossible.

Still, it's better if ColumnText doesn't make assumptions about the layout of a metadata page, in case it changes.