Refactoring: Make Sas7bdatPageLayout the source of the page size instead of Sas7bdatExporter.
Conceptually, it makes sense that the class that is responsible for laying out subheaders on a page is also responsible for determining how large the page should be.  Importantly, there is a minimum page size and this design means that Sas7bdatPageLayout can adhere to that constraint itself, instead of relying on the caller to do so.