Allow dataset names to be up to 64 bytes.
According to sas7bdat.pdf, this is how long the field is within the sas7bdat header.  It's possible that sas7bdat.pdf is wrong and the authors mistook a blank field after it as padding.  Until proof of the contrary, I'm going to assume that sas7bdat.pdf is correct.