Throw an exception from Sas7bdatExporter.close() when too few observations are written.
I had incorrectly thought that, in the context of a try-with-resources pattern, throwing an exception from close() would supersede any exception that was thrown by the body of the try block.  Based on this misconception, I thought close() would always throw an exception because the real exception would prevent the caller from writing all observations, and that close's exception would replace an actionable exception with an uninformative one.

In truth, the JVM sensibly suppresses the exception from the close(), so it doesn't mask the original exception.  Therefore, it is safe to throw an exception from close() to indicate an incorrect usage of the API.