Refactoring: Extract VariableSizeSubheader subclass.
This was done to make an internal layout pattern for some of the subheaders apparent within the Java class.  This also removes some duplicated code.

I had originally wanted VariableSizeSubheader to completely encapsulate the envelope that it writes, so that when a derived class writes at offset 0, it would be just after the "size of data" field, not the signature.  However, this made calculating the offsets too complex and made it harder to compare the implementation with the SAS7BDAT.PDF specification.