Change "data set" to "dataset".
After discussing this with the rest of Xapps, we felt that have two words made some uses unclear when "set" could be interpreted as a verb instead of a noun.  For example, "dataSetLabel()" looks like you're setting a label on data, not setting the dataset's label.

We also decided to be consistent across all uses, in documentation and code.