Stop checking the ArrayIndexOutOfBoundsException message.
Although it's a good practice to check the message to ensure that the test failed for the right reason, when I run the tests from the command line, the message is null.  This makes me think that there's a JVM option that enables the message.  To make the tests more reliable, I've removed this check.