Extend test-random-sas7bdat.groovy to test randomly generated Long and LocalDate.
Because a Long is represented as a floating point value in a SAS7BDAT, this required rewriting the logic which checks random.csv.  Instead of computing an identical string for each file (which is too hard due to the loss of precision), it compares each value in a type-sensitive way.  Number types are rounded before being compared.