Print the text fields from Row Size after all metadata is parsed.
It's not possible to print them when Row Size is parsed and still do a single pass on the file because the Column Text hasn't been parsed at that point.

This also includes some refactoring:

1) Extract getDisplayName for getting text from ColumnText and formatting it for display.
This reduces code duplication

2) Make PageReader.printSubheaderField* return the field.
This avoids having to invoke PageReader.read* with the same offsets, which precludes bugs where the offsets used for display are different from the ones used to get the value.